Add `Name` class
