+++ datacommons_client/models/node.py
@@ -2,6 +2,8 @@ from dataclasses import dataclass
 from dataclasses import field
 from typing import Any, Dict, List, Optional, TypeAlias
 
+from datacommons_client.models.base import SerializableMixin
+
 NextToken: TypeAlias = Optional[str]
 NodeDCID: TypeAlias = str
 ArcLabel: TypeAlias = str
@@ -10,7 +12,7 @@ PropertyList: TypeAlias = list[Property]
 
 
 @dataclass
-class Node:
+class Node(SerializableMixin):
   """Represents an individual node in the Data Commons knowledge graph.
 
     Attributes:
@@ -36,6 +38,21 @@ class Node:
         types=json_data.get("types"),
         value=json_data.get("value"),
     )
+
+
+@dataclass
+class Name(SerializableMixin):
+  """Represents a name associated with an Entity (node).
+
+    Attributes:
+        value: The name of the Entity
+        language: The language of the name
+        property: The property used to get the name
+    """
+
+  value: str
+  language: str
+  property: str
 
 
 @dataclass
